// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * A signed eight-byte integer. The upper big integer values are greater than the
   * max value for a JavaScript number. Therefore all big integers will be output as
   * strings and not numbers.
   */
  BigInt: { input: any; output: any; }
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: { input: any; output: any; }
  /**
   * A point in time as described by the [ISO
   * 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone.
   */
  Datetime: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export type MsgEvent = Node & {
  __typename?: 'MsgEvent';
  blockHeight: Scalars['BigInt']['output'];
  chainNum: Scalars['Int']['output'];
  /** Reads a single `Msg` that is related to this `MsgEvent`. */
  msgByChainNumAndBlockHeightAndTxIdxAndMsgIdx?: Maybe<Msg>;
  msgIdx: Scalars['Int']['output'];
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']['output'];
  /** Reads and enables pagination through a set of `Proposal`. */
  proposalsByChainNumAndBlockHeightAndTxIdxAndMsgIdxAndType: ProposalsConnection;
  /** Reads and enables pagination through a set of `Retirement`. */
  retirementsByChainNumAndBlockHeightAndTxIdxAndMsgIdxAndType: RetirementsConnection;
  txIdx: Scalars['Int']['output'];
  type: Scalars['String']['output'];
};
// ...
export type Node = {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']['output'];
};
// ...
export type Proposal = Node & {
  __typename?: 'Proposal';
  blockHeight: Scalars['BigInt']['output'];
  chainNum: Scalars['Int']['output'];
  executorResult: Scalars['String']['output'];
  finalTallyResult: Scalars['JSON']['output'];
  groupPolicyAddress: Scalars['String']['output'];
  groupPolicyVersion: Scalars['BigInt']['output'];
  groupVersion: Scalars['BigInt']['output'];
  messages: Scalars['JSON']['output'];
  metadata: Scalars['String']['output'];
  /** Reads a single `MsgEvent` that is related to this `Proposal`. */
  msgEventByChainNumAndBlockHeightAndTxIdxAndMsgIdxAndType?: Maybe<MsgEvent>;
  msgIdx: Scalars['Int']['output'];
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']['output'];
  proposalId: Scalars['BigInt']['output'];
  proposers: Array<Maybe<Scalars['String']['output']>>;
  status: Scalars['String']['output'];
  submitTime?: Maybe<Scalars['Datetime']['output']>;
  timestamp?: Maybe<Scalars['Datetime']['output']>;
  txHash: Scalars['String']['output'];
  txIdx: Scalars['Int']['output'];
  type: Scalars['String']['output'];
  votingPeriodEnd: Scalars['Datetime']['output'];
};
// ...
export type ProposalItemFragment = { __typename?: 'Proposal', type: string, blockHeight: any, txIdx: number, msgIdx: number, chainNum: number, timestamp?: any | null, txHash: string, status: string, groupPolicyAddress: string, groupPolicyVersion: any, metadata: string, proposers: Array<string | null>, submitTime?: any | null, groupVersion: any, finalTallyResult: any, votingPeriodEnd: any, executorResult: string, messages: any, id: any } & { ' $fragmentName'?: 'ProposalItemFragment' };
// ...
export const ProposalItemFragmentDoc = {"kind":"Document","definitions":[{"kind":"FragmentDefinition","name":{"kind":"Name","value":"ProposalItem"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Proposal"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"blockHeight"}},{"kind":"Field","name":{"kind":"Name","value":"txIdx"}},{"kind":"Field","name":{"kind":"Name","value":"msgIdx"}},{"kind":"Field","name":{"kind":"Name","value":"chainNum"}},{"kind":"Field","name":{"kind":"Name","value":"timestamp"}},{"kind":"Field","name":{"kind":"Name","value":"txHash"}},{"kind":"Field","alias":{"kind":"Name","value":"id"},"name":{"kind":"Name","value":"proposalId"}},{"kind":"Field","name":{"kind":"Name","value":"status"}},{"kind":"Field","name":{"kind":"Name","value":"groupPolicyAddress"}},{"kind":"Field","name":{"kind":"Name","value":"groupPolicyVersion"}},{"kind":"Field","name":{"kind":"Name","value":"metadata"}},{"kind":"Field","name":{"kind":"Name","value":"proposers"}},{"kind":"Field","name":{"kind":"Name","value":"submitTime"}},{"kind":"Field","name":{"kind":"Name","value":"groupVersion"}},{"kind":"Field","name":{"kind":"Name","value":"groupPolicyAddress"}},{"kind":"Field","name":{"kind":"Name","value":"finalTallyResult"}},{"kind":"Field","name":{"kind":"Name","value":"votingPeriodEnd"}},{"kind":"Field","name":{"kind":"Name","value":"executorResult"}},{"kind":"Field","name":{"kind":"Name","value":"messages"}}]}}]} as unknown as DocumentNode<ProposalItemFragment, unknown>;